+++ sepes/_src/tree_index.py
@@ -162,8 +162,12 @@ class BaseKey(abc.ABC):
     def __eq__(self, entry: KeyEntry) -> bool:
         pass
 
+    broadcastable: bool = False
+
 
 class IntKey(BaseKey):
+    """Match a leaf with a given index."""
+
     def __init__(self, idx: int) -> None:
         self.idx = idx
 
@@ -184,6 +188,8 @@ class IntKey(BaseKey):
 
 
 class NameKey(BaseKey):
+    """Match a leaf with a given key."""
+
     def __init__(self, name: str) -> None:
         self.name = name
 
@@ -209,6 +215,8 @@ class NameKey(BaseKey):
 
 class EllipsisKey(BaseKey):
     """Match all leaves."""
+
+    broadcastable = True
 
     def __init__(self, _):
         del _
@@ -288,13 +296,13 @@ BaseKey.def_alias = indexer_dispatcher.register
 
 _INVALID_INDEXER = """\
 Indexing with {indexer} is not implemented, supported indexing types are:
-- `str` for mapping keys or class attributes.
-- `int` for positional indexing for sequences.
-- `...` to select all leaves.
-- Boolean mask of a compatible structure as the pytree.
-- `re.Pattern` to index all keys matching a regex pattern.
-- Instance of `BaseKey` with custom logic to index a pytree.
-- `tuple` of the above types to match multiple leaves at the same level.
+  - `str` for mapping keys or class attributes.
+  - `int` for positional indexing for sequences.
+  - `...` to select all leaves.
+  - Boolean mask of a compatible structure as the pytree.
+  - `re.Pattern` to index all keys matching a regex pattern.
+  - Instance of `BaseKey` with custom logic to index a pytree.
+  - `tuple` of the above types to match multiple leaves at the same level.
 """
 
 _NO_LEAF_MATCH = """\
@@ -307,32 +315,81 @@ Check the following:
 """
 
 
-def _generate_path_mask(
-    tree: PyTree,
-    where: tuple[BaseKey, ...],
-    is_leaf: Callable[[Any], None] | None = None,
-) -> PyTree:
-    # generate a boolean mask for `where` path in `tree`
-    # where path is a tuple of indices or keys, for example
-    # where=("a",) wil set all leaves of `tree` with key "a" to True and
-    # all other leaves to False
-    match = False
+_, atomicdef = treelib.tree_flatten(0)
 
-    def map_func(path, _: Any):
-        if len(where) > len(path):
-            # path is shorter than `where` path. for example
-            # where=("a", "b") and the current path is ("a",) then
-            # the current path is not a match
+
+def one_level_tree_path_map(func, tree, *, is_leaf=None):
+    # apply func to the immediate children of tree
+    def is_leaf_func(node) -> bool:
+        # enable immediate children only
+        if is_leaf and is_leaf(node) is True:
+            return True
+        if id(node) == id(tree):
             return False
-        for wi, ki in zip(where, path):
-            if not (wi == ki):
-                return False
+        return True
 
-        nonlocal match
-        match = True
-        return match
+    return treelib.tree_path_map(func, tree, is_leaf=is_leaf_func)
 
-    mask = treelib.tree_path_map(map_func, tree, is_leaf=is_leaf)
+
+def generate_path_mask(tree, where: tuple[BaseKey, ...], *, is_leaf=None):
+    # given a pytree `tree` and a `where` path, that is composed of keys
+    # generate a boolean mask that will be eventually used to with `tree_map`
+    # to mark the leaves at the specified location.
+    # for example for a tree = [[1, 2], 3, 4] and where = [0][1] then
+    # generate [[False, True], False, False] mask
+    match: bool = False
+
+    if any(mask.broadcastable for mask in where):
+        # should the selected subtree be broadcasted to the full tree
+        # e.g. tree = [[1, 2], 3, 4] and where = [0], then
+        # broadcast with True will be [[True, True], False, False]
+        # and without broadcast will be [True, False, False]
+        # the difference is that with broadcast the user defined value will
+        # be broadcasted to the full subtree, for example if the user defined
+        # value is 100 then the result will be [[100, 100], 3, 4]
+        # and without broadcast the result will be [100, 3, 4]
+
+        def bool_tree(value: bool, tree: Any, *, is_leaf=None):
+            leaves, treedef = treelib.tree_flatten(tree, is_leaf=is_leaf)
+            return treelib.tree_unflatten(treedef, [value] * len(leaves))
+
+        true_tree = ft.partial(bool_tree, True)
+        false_tree = ft.partial(bool_tree, False)
+
+    else:
+        true_tree = lambda _, *, is_leaf: True
+        false_tree = lambda _, *, is_leaf: False
+
+    def path_map_func(path, leaf):
+        # this function is evaluated depth-first manner via
+        # `one_level_tree_path_map` to generate a boolean mask
+        nonlocal match, where
+
+        if len(path) < len(where):
+            (ki, *kr), (pi, *_) = where, path
+            if ki == pi:
+                # traverse depth-first with reduced `where` path
+                where = kr
+                out_tree = one_level_tree_path_map(path_map_func, leaf, is_leaf=is_leaf)
+                # restore `where` path for the traversal of the next sibling
+                where = (ki, *kr)
+                return out_tree
+            # exit early if the current entry in `where` path does not match
+            # the current entry in `path` path
+            return false_tree(leaf, is_leaf=is_leaf)
+
+        if len(path) == len(where):
+            for pi, ki in zip(path, where):
+                if pi != ki:
+                    return false_tree(leaf, is_leaf=is_leaf)
+            match = True
+            return true_tree(leaf, is_leaf=is_leaf)
+        # length of the path is greater than the length of `where` path
+        # e.g. where = [0][1] and path = [0][1][2] then the current path
+        # is not a match
+        return false_tree(leaf, is_leaf=is_leaf)
+
+    mask = one_level_tree_path_map(path_map_func, tree)
 
     if not match:
         path_leaf, _ = treelib.tree_path_flatten(tree, is_leaf=is_leaf)
@@ -342,24 +399,25 @@ def _generate_path_mask(
     return mask
 
 
-def _combine_bool_leaves(*leaves):
-    verdict = True
-    for leaf in leaves:
-        verdict &= leaf
-    return verdict
-
-
-def _is_bool_leaf(leaf: Any) -> bool:
-    if isinstance(leaf, arraylib.ndarray):
-        return arraylib.is_bool(leaf)
-    return isinstance(leaf, bool)
-
-
-def _resolve_where(
+def resolve_where(
     where: tuple[Any, ...],  # type: ignore
     tree: T,
     is_leaf: Callable[[Any], None] | None = None,
-) -> T | None:
+):
+    def combine_bool_leaves(*leaves):
+        # given a list of boolean leaves, combine them using `and`
+        # this is used to combine multiple boolean masks resulting from
+        # either path mask or boolean mask
+        verdict = True
+        for leaf in leaves:
+            verdict &= leaf
+        return verdict
+
+    def is_bool_leaf(leaf: Any) -> bool:
+        if isinstance(leaf, arraylib.ndarray):
+            return arraylib.is_bool(leaf)
+        return isinstance(leaf, bool)
+
     # given a pytree `tree` and a `where` path, that is composed of keys or
     # boolean masks, generate a boolean mask that will be eventually used to
     # with `tree_map` to select the leaves at the specified location.
@@ -378,7 +436,7 @@ def _resolve_where(
         # used with `is_leaf` argument of any `tree_*` function
         leaves, treedef = treelib.tree_flatten(x)
 
-        if all(map(_is_bool_leaf, leaves)):
+        if all(map(is_bool_leaf, leaves)):
             # if all leaves are boolean then this is maybe a boolean mask.
             # Maybe because the boolean mask can be a valid pytree of same structure
             # as the pytree to be indexed or _compatible_ structure.
@@ -403,10 +461,13 @@ def _resolve_where(
 
         if type(x) is tuple and seen_tuple is False:
             # e.g. `at[1,2,3]` but not `at[1,(2,3)]``
+            # i.e. inside `__getitem__` mutliple entries are transformed to a tuple
             seen_tuple = True
             return False
 
         # not a container of other keys or a pytree of same structure
+        # emit a descriptive error message to the user by pointing to the
+        # the available keys in the pytree.
         raise NotImplementedError(_INVALID_INDEXER.format(indexer=x, treedef=treedef0))
 
     for level_keys in where:
@@ -422,11 +483,11 @@ def _resolve_where(
         seen_tuple = False
 
     if path_masks:
-        mask = _generate_path_mask(tree, path_masks, is_leaf=is_leaf)
+        mask = generate_path_mask(tree, path_masks, is_leaf=is_leaf)
 
     if bool_masks:
         all_masks = [mask, *bool_masks] if mask else bool_masks
-        mask = treelib.tree_map(_combine_bool_leaves, *all_masks)
+        mask = treelib.tree_map(combine_bool_leaves, *all_masks)
 
     return mask
 
@@ -447,16 +508,18 @@ class AtIndexer:
             - a tuple of the above to match multiple keys at the same level.
 
     Example:
-        >>> # use `AtIndexer` on a pytree (e.g. dict,list,tuple,esp.)
         >>> import jax
         >>> import sepes as sp
         >>> tree = {"level1_0": {"level2_0": 100, "level2_1": 200}, "level1_1": 300}
         >>> indexer = sp.AtIndexer(tree)
+        <BLANKLINE>
         >>> indexer["level1_0"]["level2_0"].get()
         {'level1_0': {'level2_0': 100, 'level2_1': None}, 'level1_1': None}
+        <BLANKLINE>
         >>> # get multiple keys at once at the same level
         >>> indexer["level1_0"]["level2_0", "level2_1"].get()
         {'level1_0': {'level2_0': 100, 'level2_1': 200}, 'level1_1': None}
+        <BLANKLINE>
         >>> # get with a mask
         >>> mask = {"level1_0": {"level2_0": True, "level2_1": False}, "level1_1": True}
         >>> indexer[mask].get()
@@ -491,9 +554,12 @@ class AtIndexer:
         self.tree = tree
         self.where = where
 
-    def __getitem__(self, where: str) -> Self:
+    def __getitem__(self, where: Any) -> Self:
         # syntax sugar for extending the current path with `where`
         # AtIndexer(tree)[`where1`][`where2`] -> AtIndexer(tree, (`where1`, `where2`))
+        # no distinction between class attribute and mapping key is made
+        # for example AtIndexer(tree)["a"]["b"] will match both
+        # tree.a.b and tree["a"]["b"] if tree is a dict or a class instance
         return type(self)(self.tree, (*self.where, where))
 
     def __repr__(self) -> str:
@@ -539,8 +605,6 @@ class AtIndexer:
             >>> tree.at['a'].get()
             Tree(a=1, b=None)
         """
-        where = _resolve_where(self.where, self.tree, is_leaf)
-        config = dict(is_leaf=is_leaf, is_parallel=is_parallel)
 
         def leaf_get(where: Any, leaf: Any):
             # support both array and non-array leaves
@@ -553,7 +617,13 @@ class AtIndexer:
             # and `None` otherwise
             return leaf if where else None
 
-        return treelib.tree_map(leaf_get, where, self.tree, **config)
+        return treelib.tree_map(
+            leaf_get,
+            resolve_where(self.where, self.tree, is_leaf),
+            self.tree,
+            is_leaf=is_leaf,
+            is_parallel=is_parallel,
+        )
 
     def set(
         self,
@@ -561,7 +631,7 @@ class AtIndexer:
         *,
         is_leaf: Callable[[Any], None] | None = None,
         is_parallel: bool | ParallelConfig = False,
-    ) -> PyTree:
+    ):
         """Set the leaf values at the specified location.
 
         Args:
@@ -597,8 +667,6 @@ class AtIndexer:
             >>> tree.at['a'].set(100)
             Tree(a=100, b=2)
         """
-        where = _resolve_where(self.where, self.tree, is_leaf)
-        config = dict(is_leaf=is_leaf, is_parallel=is_parallel)
 
         def leaf_set(where: Any, leaf: Any, set_value: Any):
             # support both array and non-array leaves
@@ -619,12 +687,22 @@ class AtIndexer:
             # to tree2 leaves if tree2 is a pytree of same structure as tree
             # instead of making each leaf of tree a copy of tree2
             # is design is similar to ``numpy`` design `np.at[...].set(Array)`
-            return treelib.tree_map(leaf_set, where, self.tree, set_value, **config)
+            return treelib.tree_map(
+                leaf_set,
+                resolve_where(self.where, self.tree, is_leaf),
+                self.tree,
+                set_value,
+                is_leaf=is_leaf,
+                is_parallel=is_parallel,
+            )
 
-        # set_value is broadcasted to tree leaves
-        # for example tree.at[where].set(1) will set all tree leaves to 1
-        leaf_set_ = lambda where, leaf: leaf_set(where, leaf, set_value)
-        return treelib.tree_map(leaf_set_, where, self.tree, **config)
+        return treelib.tree_map(
+            ft.partial(leaf_set, set_value=set_value),
+            resolve_where(self.where, self.tree, is_leaf),
+            self.tree,
+            is_leaf=is_leaf,
+            is_parallel=is_parallel,
+        )
 
     def apply(
         self,
@@ -632,7 +710,7 @@ class AtIndexer:
         *,
         is_leaf: Callable[[Any], None] | None = None,
         is_parallel: bool | ParallelConfig = False,
-    ) -> PyTree:
+    ):
         """Apply a function to the leaf values at the specified location.
 
         Args:
@@ -675,8 +753,6 @@ class AtIndexer:
             >>> indexer = sp.AtIndexer({"lenna": "lenna.png", "baboon": "baboon.png"})
             >>> images = indexer[...].apply(imread, parallel=dict(max_workers=2))  # doctest: +SKIP
         """
-        where = _resolve_where(self.where, self.tree, is_leaf)
-        config = dict(is_leaf=is_leaf, is_parallel=is_parallel)
 
         def leaf_apply(where: Any, leaf: Any):
             # same as `leaf_set` but with `func` applied to the leaf
@@ -687,7 +763,13 @@ class AtIndexer:
                 return arraylib.where(where, func(leaf), leaf)
             return func(leaf) if where else leaf
 
-        return treelib.tree_map(leaf_apply, where, self.tree, **config)
+        return treelib.tree_map(
+            leaf_apply,
+            resolve_where(self.where, self.tree, is_leaf),
+            self.tree,
+            is_leaf=is_leaf,
+            is_parallel=is_parallel,
+        )
 
     def scan(
         self,
@@ -748,7 +830,6 @@ class AtIndexer:
             them with final state. While ``reduce`` applies a binary ``func`` to the
             leaf values while carrying a state and returning a single value.
         """
-        where = _resolve_where(self.where, self.tree, is_leaf)
 
         running_state = state
 
@@ -762,8 +843,13 @@ class AtIndexer:
                 return arraylib.where(where, stateless_func(leaf), leaf)
             return stateless_func(leaf) if where else leaf
 
-        out = treelib.tree_map(leaf_apply, where, self.tree, is_leaf=is_leaf)
-        return out, running_state
+        out_tree = treelib.tree_map(
+            leaf_apply,
+            resolve_where(self.where, self.tree, is_leaf),
+            self.tree,
+            is_leaf=is_leaf,
+        )
+        return out_tree, running_state
 
     def reduce(
         self,